Move more examples
